data: More cross-backend homogeneity
