+++ devito/data.py
